Module 7
Module 7 catch me if you can activity

+++ KristinaBootcamp/src/com/bootcampexcercise/module2/activity/OperatorActivity.java
@@ -4,7 +4,7 @@ public class OperatorActivity {
     public static void main(String args[]) {
         int x = 5;
         int y = 4;
-        int sum = x+y;
+        double sum = x+y;
         System.out.println("Sum of number is " +sum);
 
     }
